Revert "[pre-commit.ci] auto fixes from pre-commit.com hooks"
This reverts commit c64d5b475e7f176436484ce19077fae571a73b1a.

+++ tests/saving/language_models/test_merge_4bit_validation.py
@@ -17,7 +17,7 @@ def formatting_prompts_func(examples):
     convos = examples["messages"]
     texts = [
         tokenizer.apply_chat_template(
-            convo, tokenize=False, add_generation_prompt=False
+            convo, tokenize = False, add_generation_prompt = False
         )
         for convo in convos
     ]
@@ -36,25 +36,25 @@ else:
     attn_implementation = "sdpa"
 
 model, tokenizer = FastLanguageModel.from_pretrained(
-    model_name="unsloth/Llama-3.1-8B-Instruct",
-    max_seq_length=2048,
-    dtype=compute_dtype,
-    load_in_4bit=True,
-    load_in_8bit=False,
-    full_finetuning=False,
-    attn_implementation=attn_implementation,
+    model_name = "unsloth/Llama-3.1-8B-Instruct",
+    max_seq_length = 2048,
+    dtype = compute_dtype,
+    load_in_4bit = True,
+    load_in_8bit = False,
+    full_finetuning = False,
+    attn_implementation = attn_implementation,
 )
 
 tokenizer = get_chat_template(
     tokenizer,
-    chat_template="llama-3.1",
+    chat_template = "llama-3.1",
 )
 
 # Load small dataset for quick training
 dataset_train = load_dataset(
-    "allenai/openassistant-guanaco-reformatted", split="train[:100]"
+    "allenai/openassistant-guanaco-reformatted", split = "train[:100]"
 )
-dataset_train = dataset_train.map(formatting_prompts_func, batched=True)
+dataset_train = dataset_train.map(formatting_prompts_func, batched = True)
 
 print("✅ Base model loaded successfully!")
 
@@ -64,8 +64,8 @@ print(f"{'='*80}")
 
 model = FastLanguageModel.get_peft_model(
     model,
-    r=16,
-    target_modules=[
+    r = 16,
+    target_modules = [
         "k_proj",
         "q_proj",
         "v_proj",
@@ -74,40 +74,40 @@ model = FastLanguageModel.get_peft_model(
         "down_proj",
         "up_proj",
     ],
-    lora_alpha=16,
-    lora_dropout=0,
-    bias="none",
-    use_gradient_checkpointing="unsloth",
-    random_state=3407,
-    use_rslora=False,
-    loftq_config=None,
+    lora_alpha = 16,
+    lora_dropout = 0,
+    bias = "none",
+    use_gradient_checkpointing = "unsloth",
+    random_state = 3407,
+    use_rslora = False,
+    loftq_config = None,
 )
 
 from unsloth import is_bfloat16_supported
 
 trainer = SFTTrainer(
-    model=model,
-    tokenizer=tokenizer,
-    train_dataset=dataset_train,
-    dataset_text_field="text",
-    max_seq_length=2048,
-    data_collator=DataCollatorForSeq2Seq(tokenizer=tokenizer),
-    dataset_num_proc=2,
-    packing=False,
-    args=TrainingArguments(
-        per_device_train_batch_size=2,
-        gradient_accumulation_steps=4,
-        warmup_ratio=0.1,
-        max_steps=10,  # Very short training for test
-        learning_rate=2e-4,
-        fp16=not is_bfloat16_supported(),
-        bf16=is_bfloat16_supported(),
-        logging_steps=5,
-        optim="adamw_8bit",
-        lr_scheduler_type="linear",
-        seed=3407,
-        output_dir="outputs",
-        report_to="none",
+    model = model,
+    tokenizer = tokenizer,
+    train_dataset = dataset_train,
+    dataset_text_field = "text",
+    max_seq_length = 2048,
+    data_collator = DataCollatorForSeq2Seq(tokenizer = tokenizer),
+    dataset_num_proc = 2,
+    packing = False,
+    args = TrainingArguments(
+        per_device_train_batch_size = 2,
+        gradient_accumulation_steps = 4,
+        warmup_ratio = 0.1,
+        max_steps = 10,  # Very short training for test
+        learning_rate = 2e-4,
+        fp16 = not is_bfloat16_supported(),
+        bf16 = is_bfloat16_supported(),
+        logging_steps = 5,
+        optim = "adamw_8bit",
+        lr_scheduler_type = "linear",
+        seed = 3407,
+        output_dir = "outputs",
+        report_to = "none",
     ),
 )
 
@@ -119,9 +119,9 @@ print("🔍 PHASE 3: Save with Forced 4bit Merge")
 print(f"{'='*80}")
 
 model.save_pretrained_merged(
-    save_directory="./test_4bit_model",
-    tokenizer=tokenizer,
-    save_method="forced_merged_4bit",
+    save_directory = "./test_4bit_model",
+    tokenizer = tokenizer,
+    save_method = "forced_merged_4bit",
 )
 
 print("✅ Model saved with forced 4bit merge!")
@@ -137,15 +137,15 @@ torch.cuda.empty_cache()
 
 # Load the 4bit merged model
 model_4bit, tokenizer_4bit = FastLanguageModel.from_pretrained(
-    model_name="./test_4bit_model",
-    max_seq_length=2048,
-    load_in_4bit=True,
-    load_in_8bit=False,
+    model_name = "./test_4bit_model",
+    max_seq_length = 2048,
+    load_in_4bit = True,
+    load_in_8bit = False,
 )
 
 tokenizer_4bit = get_chat_template(
     tokenizer_4bit,
-    chat_template="llama-3.1",
+    chat_template = "llama-3.1",
 )
 
 print("✅ 4bit model loaded successfully!")
@@ -153,8 +153,8 @@ print("✅ 4bit model loaded successfully!")
 # Add LoRA adapters to the 4bit model
 model_4bit = FastLanguageModel.get_peft_model(
     model_4bit,
-    r=16,
-    target_modules=[
+    r = 16,
+    target_modules = [
         "k_proj",
         "q_proj",
         "v_proj",
@@ -163,39 +163,39 @@ model_4bit = FastLanguageModel.get_peft_model(
         "down_proj",
         "up_proj",
     ],
-    lora_alpha=16,
-    lora_dropout=0,
-    bias="none",
-    use_gradient_checkpointing="unsloth",
-    random_state=3407,
-    use_rslora=False,
-    loftq_config=None,
+    lora_alpha = 16,
+    lora_dropout = 0,
+    bias = "none",
+    use_gradient_checkpointing = "unsloth",
+    random_state = 3407,
+    use_rslora = False,
+    loftq_config = None,
 )
 
 # Second fine-tuning
 trainer_4bit = SFTTrainer(
-    model=model_4bit,
-    tokenizer=tokenizer_4bit,
-    train_dataset=dataset_train,
-    dataset_text_field="text",
-    max_seq_length=2048,
-    data_collator=DataCollatorForSeq2Seq(tokenizer=tokenizer_4bit),
-    dataset_num_proc=2,
-    packing=False,
-    args=TrainingArguments(
-        per_device_train_batch_size=2,
-        gradient_accumulation_steps=4,
-        warmup_ratio=0.1,
-        max_steps=10,  # Very short training for test
-        learning_rate=2e-4,
-        fp16=not is_bfloat16_supported(),
-        bf16=is_bfloat16_supported(),
-        logging_steps=5,
-        optim="adamw_8bit",
-        lr_scheduler_type="linear",
-        seed=3407,
-        output_dir="outputs_4bit",
-        report_to="none",
+    model = model_4bit,
+    tokenizer = tokenizer_4bit,
+    train_dataset = dataset_train,
+    dataset_text_field = "text",
+    max_seq_length = 2048,
+    data_collator = DataCollatorForSeq2Seq(tokenizer = tokenizer_4bit),
+    dataset_num_proc = 2,
+    packing = False,
+    args = TrainingArguments(
+        per_device_train_batch_size = 2,
+        gradient_accumulation_steps = 4,
+        warmup_ratio = 0.1,
+        max_steps = 10,  # Very short training for test
+        learning_rate = 2e-4,
+        fp16 = not is_bfloat16_supported(),
+        bf16 = is_bfloat16_supported(),
+        logging_steps = 5,
+        optim = "adamw_8bit",
+        lr_scheduler_type = "linear",
+        seed = 3407,
+        output_dir = "outputs_4bit",
+        report_to = "none",
     ),
 )
 
@@ -208,8 +208,8 @@ print(f"{'='*80}")
 
 try:
     model_4bit.save_pretrained_merged(
-        save_directory="./test_should_fail",
-        tokenizer=tokenizer_4bit,
+        save_directory = "./test_should_fail",
+        tokenizer = tokenizer_4bit,
         # No save_method specified, should default to regular merge
     )
     assert False, "Expected TypeError but merge succeeded!"
@@ -225,9 +225,9 @@ print(f"{'='*80}")
 
 try:
     model_4bit.save_pretrained_merged(
-        save_directory="./test_4bit_second",
-        tokenizer=tokenizer_4bit,
-        save_method="forced_merged_4bit",
+        save_directory = "./test_4bit_second",
+        tokenizer = tokenizer_4bit,
+        save_method = "forced_merged_4bit",
     )
     print("✅ Successfully saved 4bit model with forced 4bit method!")
 except Exception as e:
